+++ source/QSAR/nBModel.C
@@ -196,7 +196,7 @@ vector<double> NBModel::calculateProbabilities(int activitiy_index, int feature_
 	}
 	uint no_features = probabilities_[0][0].Ncols();
 	uint no_classes = probabilities_[0].size();
-	if(activitiy_index>=(int)probabilities_.size() || feature_index>=no_features || activitiy_index<0 || feature_index<0)
+	if(activitiy_index>=(int)probabilities_.size() || feature_index>=(int)no_features || activitiy_index<0 || feature_index<0)
 	{
 		throw Exception::InconsistentUsage(__FILE__,__LINE__,"Index of bound for parameters given to SNBModel::calculateProbability() !");
 	}
